doc(debugging): add VS Code debugging guide
This is a basic and first step towards debugging pytest with VS Code. The motivation is that I was not having much success using `--pdb` flags and `pdb.set_trace()` triggers. Instead of allowing single step execution as I had hoped, *pytest* would perform all of its test run without the supervision of the debugger.

Does that behaviour have something to do with how processes are handled?