experiment: change destination of link
The current link shows a 404 message in the diagram frame.
This is a new error and unexpected error. Normally, it is a "Content is blocked" error when linking to GitHub pages. However, the url doesn't point to an existing page. This commit changes the url so that it points at an existing link.